Add support for key-values and csv data
Parameters can be supplied in the input file as key-value pairs.
Everything except file type has a default and need not be specified.
Key values can be specified in any order.
There is a separate function for reading point clouds in csv that does the same thing as the original read_point_cloud function.
Functionality and support is retained for original data files.