Add new function flag
The --function flag is not available on the command line.
When in a file, it must be followed by a list of values.
If not supplied, data is assumed to not have a function.
All old file types and functionalities are still supported.
The parse_metric_old() function is added to parse old metric input files.
The new metric type input files can be parsed in parse_args().