Separate InputManager into itself and DataReader
InputManager was reading in the entire file and performing calculations.
Now it only reads in the flags that are supplied before the data.
DataReader reads in the data and performs the rest of the calculations.
This allows InputManager to be used in the GUI section without dependencies.
Other files are modified appropriately.