Add "inf" option for max distance
User may specify a number or "inf" for infinity.
If the box is active, and the mouse hovers over it, a tool tip will show up.
InputParameters now has a variable called md_string which stores the string value of max distance.
This is to accomodate the input as a text box instead of spin box.